refactor: Repo, User Api 주석 추가
