+++ Moyoy-Infra/src/main/java/com/moyoy/infra/database/mysql/github/OAuthTokenJDBCRepository.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import com.moyoy.infra.external.github.helper.AESEncryptor;
+import com.moyoy.infra.database.mysql.support.AESEncryptor;
 
 @Repository
 @RequiredArgsConstructor
