+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/github/follow/GithubFollowApi.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.moyoy.infra.external.github.user.GithubUserResponse;
-import com.moyoy.infra.external.github.support.config.GithubFeignConfig;
+import com.moyoy.infra.external.github.user.dto.GithubUserResponse;
+import com.moyoy.infra.external.github.support.GithubFeignConfig;
 
 import feign.Response;
