+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/github/repo/dto/GithubRepoContributorsResponse.java
@@ -1,4 +1,4 @@
-package com.moyoy.infra.external.github.repo;
+package com.moyoy.infra.external.github.repo.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
