+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/github/user/dto/GithubUserResponse.java
@@ -1,4 +1,4 @@
-package com.moyoy.infra.external.github.user;
+package com.moyoy.infra.external.github.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
